PEST testing for cupcakes, setup PEST testing for promocode

+++ routes/api.php
@@ -3,6 +3,7 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CartController;
+use App\Http\Middleware\EnsureUserIsAdmin;
 use App\Http\Controllers\CupcakeController;
 
 Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
@@ -16,9 +17,9 @@ Route::get('/test', function(){
 // Cupcake
 Route::get('/cupcake', [CupcakeController::class, 'getCupcakes']);
 Route::get('/cupcake/{id}',  [CupcakeController::class, 'getCupcake']);
-Route::put('/cupcake/{id}',  [CupcakeController::class, 'updateCupcake']);
-Route::post('/cupcake', [CupcakeController::class, 'createCupcake']);
-Route::delete('/cupcake/{id}', [CupcakeController::class, 'deleteCupcake']);
+Route::put('/cupcake/{id}',  [CupcakeController::class, 'updateCupcake'])->middleware(EnsureUserIsAdmin::class);
+Route::post('/cupcake', [CupcakeController::class, 'createCupcake'])->middleware(EnsureUserIsAdmin::class);
+Route::delete('/cupcake/{id}', [CupcakeController::class, 'deleteCupcake'])->middleware(EnsureUserIsAdmin::class);
 
 // Cart
 Route::post('/cart', [CartController::class, 'upsert']);
